fixed an issue in chunk making

+++ code/pre_process.py
@@ -129,7 +129,7 @@ def make_chunks(input_data, subjects):
             subject_data = input_data[f"subject{subject}"]
             np.concatenate((data, subject_data[:1278]), axis=0)
             
-    chunks = np.zeros((1278*32/6, 6, 32, 128))
+    chunks = np.zeros((int(1278*32/6), 6, 32, 128))
 
     for trail in range (chunks.shape[0]):
         chunks[trail, :, :, :] = data[trail*6:(trail+1)*6, :, :]
